+++ lib/prompt/io/text.ex
@@ -31,6 +31,8 @@ defmodule Prompt.IO.Text do
   defimpl Prompt.IO do
     @spec display(Prompt.IO.Text.t()) :: Prompt.IO.Text.t()
     def display(txt) do
+      _ = Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :cooked})
+
       [
         :reset,
         background_color(txt),